Updates made via Claude code so that it works with modern iOS
  Project Configuration Changes (project.pbxproj)

  1. Updated iOS Deployment Target: 8.0 → 12.0 (minimum supported by current Xcode)
  2. Fixed deprecated build settings:
    - ALWAYS_SEARCH_USER_PATHS: YES → NO
    - Removed deprecated VALID_ARCHS setting
  3. Updated Header Search Paths to properly resolve Box2D includes after disabling user path search
  4. Removed duplicate README.md files from Copy Bundle Resources phase (JSONKit and TestFlightSDK)
  5. Added compiler flags:
    - OTHER_CPLUSPLUSFLAGS: Added -Wno-c++11-narrowing to suppress narrowing conversion errors in legacy code
    - GCC_PREPROCESSOR_DEFINITIONS: Added GLES_SILENCE_DEPRECATION to quiet OpenGL ES deprecation warnings

  Code Fixes

  1. SettingsManager.mm: Fixed nil return values for primitive types (bool, int, double) which are no longer allowed in modern compilers
  2. ccTypes.h (cocos2d): Added explicit casts in ccc4BFromccc4F() to fix C++11 narrowing errors
  3. JSONKit.m: Replaced deprecated direct isa assignment with object_setClass() calls
  4. Initialized git submodule: Fetched the JSONKit library which was an empty submodule

+++ Penguin Rescue/libs/cocos2d/ccTypes.h
@@ -141,7 +141,7 @@ static inline BOOL ccc4FEqual(ccColor4F a, ccColor4F b)
 	
 static inline ccColor4B ccc4BFromccc4F(ccColor4F c)
 {
-	return (ccColor4B){c.r*255.f, c.g*255.f, c.b*255.f, c.a*255.f};
+	return (ccColor4B){(GLubyte)(c.r*255.f), (GLubyte)(c.g*255.f), (GLubyte)(c.b*255.f), (GLubyte)(c.a*255.f)};
 }
